ScheduleRestartInfo: put in separate file
